requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
langchain-huggingface

groq>=0.4.1
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pathlib>=1.0.1
logging>=0.4.9.6
pickle-mixin>=1.0.2
langchain>=0.0.350
langchain-community>=0.0.10
tiktoken>=0.5.0
faiss-cpu>=1.7.4

# For document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# For web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0